reorganization of the project
