+++ src/blizzApiWrapper/apiAccess.js
@@ -1,0 +1,4 @@
+export default {
+  clientId: `e4b0c1cf4f14416cb1f68a417f53e6cc`,
+  clientSecret: `DtLcZEraCXzSmEoTU7GC8oLJZs4QPB1O`,
+};
