+++ src/blizzApiWrapper/toDo.txt
@@ -1,0 +1,2 @@
+find out how it works
+blizzApi
