+++ src/components/Heartstone/styles/Result.css
